VISUALIZATION: argparse in visualize_msg.py

+++ visualize_msg.py
@@ -6,15 +6,18 @@ import matplotlib.pyplot as plt
 import os
 import sys
 import shutil
+import argparse
 
 # generate plots on ax given an adjacency dictionary
+
 
 class MessageGraph:
     def __init__(self, all_nodes, indices, lim):
         self.all_nodes = all_nodes
         self.indices = indices
         self.lim = lim
-    
+        self.max_node_length = max(map(lambda i: len(str(i)), all_nodes))
+
     def get_deg(self, node):
         node = int(node)
         all_nodes, indices = self.all_nodes, self.indices
@@ -27,40 +30,39 @@ class MessageGraph:
         else:
             deg = all_nodes.index(node)/(len(all_nodes)) * 2 * np.pi
         return deg
-        
-    def plot(self, ax, adj, msgs, holds):
+
+    def plot(self, ax, adj, msgs, holds, truncate):
         lim = self.lim
         ax.clear()
         ax.axis('off')
         ax.set_xlim(left=-lim, right=lim)
         ax.set_ylim(top=lim, bottom=-lim)
-        self.plot_adj(adj, ax)
+        self.plot_adj(adj, ax, truncate)
         self.plot_msg(msgs, ax)
         self.plot_holds(holds, ax)
 
     def plot_msg(self, msgs, ax):
         for a, b in msgs:
             degs = list(map(self.get_deg, (a, b)))
-            
+
             x1, x2 = np.cos(degs)
             y1, y2 = np.sin(degs)
 
             ax.annotate("",
-                xy=(x2, y2), xycoords='data',
-                xytext=(x1, y1), textcoords='data',
-                arrowprops=dict(arrowstyle="->",
-                                mutation_scale=15,
-                                lw=2,
-                                connectionstyle="arc3",
-                                color="tab:green"),
-                )
+                        xy=(x2, y2), xycoords='data',
+                        xytext=(x1, y1), textcoords='data',
+                        arrowprops=dict(arrowstyle="->",
+                                        mutation_scale=15,
+                                        lw=2,
+                                        connectionstyle="arc3",
+                                        color="tab:green"),
+                        )
 
     def plot_holds(self, holds, ax):
         degs = list(map(self.get_deg, holds))
         ax.scatter(np.cos(degs), np.sin(degs), zorder=10)
-        
 
-    def plot_adj(self, adj, ax):
+    def plot_adj(self, adj, ax, truncate):
         # retrieve the edges in the adjacency dictionary
         edges = [(int(k), int(v)) for k, l in adj.items() for v in l]
 
@@ -84,14 +86,30 @@ class MessageGraph:
             c, s = np.cos(deg)*scale, np.sin(deg)*scale
 
             # label the coordinate
-            ax.text(c, s, node[:4],
+            label = node if truncate == - \
+                1 else node[:truncate + len(node) - self.max_node_length]
+            ax.text(c, s, label,
                     horizontalalignment='center',
                     verticalalignment='center')
 
+
 if __name__ == '__main__':
-    # read the adjacency matrix JSON file
+
+    parser = argparse.ArgumentParser(description='visualize chord graph')
+    parser.add_argument('output', help='output file/folder')
+    parser.add_argument('--input', '-i', default='data.json',
+                        help='input json file')
+    parser.add_argument('--folder', '-f', action='store_true',
+                        help='output as folder')
+    parser.add_argument('--truncate', '-t', nargs='?', type=int, const=4,
+                        default=-1, help='truncate node names')
+    parser.add_argument('--last', '-l', action='store_true',
+                        help='only plot last snapshot')
+    args = parser.parse_args()
+
+ # read the adjacency matrix JSON file
     figs_path = os.path.join(os.getcwd(), 'figs')
-    in_file = sys.argv[1] if len(sys.argv) == 3 else 'data.json'
+    in_file = args.input
     with open(os.path.join(os.getcwd(), in_file), 'r') as f:
         snapshots = json.load(f)
 
@@ -106,19 +124,22 @@ if __name__ == '__main__':
 
     graph = MessageGraph(all_nodes, indices, lim)
 
-    figs_name = sys.argv[-1]
+    if args.folder:
+        export_path = os.path.join(figs_path, args.output)
+        if os.path.exists(export_path):
+            shutil.rmtree(export_path)
+        os.mkdir(export_path)
+        def filename(i): return '%d.png' % i
+    else:
+        export_path = figs_path
+        def filename(i): return '%s-%i.png' % (args.output, i)
 
-    export_path = os.path.join(figs_path, figs_name)
-
-    if os.path.exists(export_path):
-        shutil.rmtree(export_path)
-    os.mkdir(export_path)
-
-    for i, t in enumerate(snapshots):
+    enum = [(len(snapshots) - 1, snapshots[-1])
+            ] if args.last else enumerate(snapshots)
+    for i, t in enum:
         # snapshots: adj, msgs, holders
         adj = t['adj']
         msgs = t['msgs']
         holders = t['holders']
-        graph.plot(ax, adj, msgs, holders)
-        plt.savefig(os.path.join(export_path, '%d.png' %
-                                 i), dpi=300)
+        graph.plot(ax, adj, msgs, holders, truncate=args.truncate)
+        plt.savefig(os.path.join(export_path, filename(i)), dpi=300)
